back/requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
drf-spectacular[sidecar]==0.29.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.0.0
